🚚 extract handle_request_validation_error from app services to utils
